bot/requirements.txt
discord.py==2.3.1
Flask==2.3.3
python-dotenv==1.0.0
pytz==2023.3
requests